Update and rename AA1.R to LR1-Logistics Regression Model.R
The main purpose of this lab report is to build a statistical model that best predicts the survival probability of Sue and Kate separately with or without having Leonardo with them during the
Titanic incident. The task was adapted from the Titanic Challenge on Kaggle(https://www.kaggle.com/c/titanic/overview).